isolating the files used from packages
MODIFIED: I didn't include all the banana plants and many of the files from the computer workstation which are default but we aren't using.

This should be most of the content used from the free packs. Some materials may not be there but if you get the free pack on your local, it should update accordingly. Fully Pack list to come.